multi collateral proxy completed

+++ contracts/interfaces/IWETH.sol
@@ -6,5 +6,5 @@ import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 interface IWETH is IERC20 {
     function deposit() external;
 
-    function withdraw(uint) external;
+    function withdraw(uint256) external;
 }
